fix negative progress bar
git-svn-id: file:///autofs/homes/012/fangq/space/git/Temp/mmc_svn@121 4848f366-3658-0410-a320-90152d9d5b07

+++ mmc/trunk/src/mcx_utils.c
@@ -280,7 +280,7 @@ int mcx_parsedebugopt(char *debugopt){
     return debuglevel;
 }
 
-void mcx_progressbar(int n, int ntotal, Config *cfg){
+void mcx_progressbar(unsigned int n, unsigned int ntotal, Config *cfg){
     unsigned int percentage, j,colwidth=79;
     static unsigned int oldmarker=0;
 
@@ -290,7 +290,7 @@ void mcx_progressbar(int n, int ntotal, Config *cfg){
     colwidth=ttys.ws_col;
 #endif
     
-    percentage=n*(colwidth-18)/ntotal;
+    percentage=(float)n*(colwidth-18)/ntotal;
 
     if(percentage != oldmarker){
         oldmarker=percentage;
